requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
blis==1.2.0
Bottleneck==1.4.2
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cuda-python==12.6.0
cudf-cu12==24.12.0
cupy-cuda12x==13.3.0
cycler==0.12.1
cymem==2.0.11
docutils==0.21.2
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastai==1.0.57
fastprogress==1.0.3
fastrlock==0.8.3
fasttext==0.9.3
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gensim==4.3.3
h11==0.14.0
hdbscan==0.8.40
idna==3.10
imagesize==1.4.1
indic-nlp-library==0.92
inltk==0.9
Jinja2==3.1.5
joblib==1.4.2
kaleido==0.2.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libcudf-cu12==24.12.0
libkvikio-cu12==24.12.1
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
Morfessor==2.0.6
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
narwhals==1.25.2
networkx==3.4.2
numba==0.61.0
numba-cuda==0.0.17.1
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.21.5
nvidia-nvcomp-cu12==4.1.0.6
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvtx==0.2.10
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
POT==0.9.5
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.3
pyarrow==18.1.0
pybind11==2.13.6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pylibcudf-cu12==24.12.0
pynndescent==0.5.13
pynvjitlink-cu12==0.5.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rmm-cu12==24.12.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
selenium==4.28.1
sentencepiece==0.2.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.1.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
srsly==2.5.1
stanza==1.10.1
sympy==1.13.1
thinc==8.3.4
threadpoolctl==3.5.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
triton==3.2.0
typer==0.15.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
